+++ spot-ml/ml_ops.sh
@@ -67,8 +67,6 @@ LDA_OUTPUT_DIR=${DSOURCE}/${FDATE}
 
 TOPIC_COUNT=20
 
-LDA_IMP="SparkLDA"
-
 nodes=${NODES[0]}
 for n in "${NODES[@]:1}" ; do nodes+=",${n}"; done
 
@@ -79,13 +77,6 @@ mkdir -p ${LPATH}
 rm -f ${LPATH}/*.{dat,beta,gamma,other,pkl} # protect the flow_scores.csv file
 
 hdfs dfs -rm -R -f ${HDFS_SCORED_CONNECTS}
-
-# Add -p <command> to execute pre MPI command.
-# Pre MPI command can be configured in /etc/spot.conf
-# In this script, after the line after --mpicmd ${MPI_CMD} add:
-# --mpiprep ${MPI_PREP_CMD}
-
-${MPI_PREP_CMD}
 
 time spark-submit --class "org.apache.spot.SuspiciousConnects" \
   --master yarn-client \
@@ -105,21 +96,10 @@ time spark-submit --class "org.apache.spot.SuspiciousConnects" \
   --input ${RAWDATA_PATH}  \
   --dupfactor ${DUPFACTOR} \
   --feedback ${FEEDBACK_PATH} \
-  --model ${LPATH}/model.dat \
-  --topicdoc ${LPATH}/final.gamma \
-  --topicword ${LPATH}/final.beta \
-  --lpath ${LPATH} \
-  --ldapath ${LDAPATH} \
-  --luser ${LUSER} \
-  --mpicmd ${MPI_CMD}  \
-  --proccount ${PROCESS_COUNT} \
   --topiccount ${TOPIC_COUNT} \
-  --nodes ${nodes} \
   --scored ${HDFS_SCORED_CONNECTS} \
-  --tempmodel ${HDFS_MODEL} \
   --threshold ${TOL} \
-  --maxresults ${MAXRESULTS}\
-  --ldaImplementation ${LDA_IMP} \
+  --maxresults ${MAXRESULTS} \
   --ldaMaxIterations 20
 
 wait
