+++ spot-ml/src/main/scala/org/apache/spot/SuspiciousConnectsArgumentParser.scala
@@ -8,27 +8,17 @@ object SuspiciousConnectsArgumentParser {
 
   case class SuspiciousConnectsConfig(analysis: String = "",
                                       inputPath: String = "",
-                                      scoresFile: String = "",
+                                      feedbackFile: String = "",
                                       duplicationFactor: Int = 1,
-                                      modelFile: String = "",
-                                      topicDocumentFile: String = "",
-                                      topicWordFile: String = "",
-                                      mpiPreparationCmd: String = "",
-                                      mpiCmd: String = "",
-                                      mpiProcessCount: String = "",
                                       topicCount: Int = 20,
-                                      localPath: String = "",
-                                      localUser: String = "",
-                                      ldaPath: String = "",
-                                      nodes: String = "",
                                       hdfsScoredConnect: String = "",
-                                      hdfsModelFile: String = "",
                                       threshold: Double = 1.0d,
                                       maxResults: Int = -1,
                                       outputDelimiter: String = "\t",
-                                      // ldaPRGSeed: Option[Long] = None,
-                                      ldaImplementation: String = "SparkLDA",
-                                      ldaMaxIterations: Int = 20)
+                                      ldaPRGSeed: Option[Long] = None,
+                                      ldaMaxIterations: Int = 20,
+                                      ldaAlpha: Double = 1.02,
+                                      ldaBeta: Double = 1.001)
 
   val parser: scopt.OptionParser[SuspiciousConnectsConfig] = new scopt.OptionParser[SuspiciousConnectsConfig]("LDA") {
 
@@ -43,64 +33,21 @@ object SuspiciousConnectsArgumentParser {
       text("HDFS path to input")
 
     opt[String]("feedback").valueName("<local file>").
-      action((x, c) => c.copy(scoresFile = x)).
+      action((x, c) => c.copy(feedbackFile = x)).
       text("the local path of the file that contains the feedback scores")
 
     opt[Int]("dupfactor").valueName("<non-negative integer>").
       action((x, c) => c.copy(duplicationFactor = x)).
       text("duplication factor controlling how to downgrade non-threatening connects from the feedback file")
 
-    opt[String]("model").required().valueName("<local file>").
-      action((x, c) => c.copy(modelFile = x)).
-      text("Model file location")
-
-    opt[String]("topicdoc").required().valueName("<local file>").
-      action((x, c) => c.copy(topicDocumentFile = x)).
-      text("final.gamma file location")
-
-    opt[String]("topicword").required().valueName("<local file>").
-      action((x, c) => c.copy(topicWordFile = x)).
-      text("final.beta file location")
-
-    opt[String]("mpiprep").valueName("<mpi command>").
-      action((x, c) => c.copy(mpiPreparationCmd = x)).
-      text("MPI preparation command")
-
-    opt[String]("mpicmd").required().valueName("<mpi command>").
-      action((x, c) => c.copy(mpiCmd = x)).
-      text("MPI command")
-
-    opt[String]("proccount").required().valueName("<mpi param>").
-      action((x, c) => c.copy(mpiProcessCount = x)).
-      text("MPI process count")
 
     opt[String]("topiccount").required().valueName("number of topics in topic model").
       action((x, c) => c.copy(topicCount = x.toInt)).
       text("topic count")
 
-    opt[String]("lpath").required().valueName("<local path>").
-      action((x, c) => c.copy(localPath = x)).
-      text("Local Path")
-
-    opt[String]("ldapath").required().valueName("<local path>").
-      action((x, c) => c.copy(ldaPath = x)).
-      text("LDA Path")
-
-    opt[String]("luser").required().valueName("<local path>").
-      action((x, c) => c.copy(localUser = x)).
-      text("Local user path")
-
-    opt[String]("nodes").required().valueName("<input param>").
-      action((x, c) => c.copy(nodes = x)).
-      text("Node list")
-
     opt[String]("scored").required().valueName("<hdfs path>").
       action((x, c) => c.copy(hdfsScoredConnect = x)).
       text("HDFS path for results")
-
-    opt[String]("tempmodel").required().valueName("<hdfs path>").
-      action((x, c) => c.copy(hdfsModelFile = x)).
-      text("HDFS path for model (temporary location)")
 
     opt[Double]("threshold").required().valueName("float64").
       action((x, c) => c.copy(threshold = x)).
@@ -110,21 +57,25 @@ object SuspiciousConnectsArgumentParser {
       action((x, c) => c.copy(maxResults = x)).
       text("number of most suspicious connections to return")
 
-    opt[String]("ldaImplementation").required().valueName("< LDAC | sparkLDA >").
-      action((x, c) => c.copy(ldaImplementation = x)).
-      text("choice of which implementation to use for LDA, options are LDAC or SparkLDA")
-
     opt[String]("delimiter").optional().valueName("character").
       action((x, c) => c.copy(outputDelimiter = x)).
       text("number of most suspicious connections to return")
 
-    /*
     opt[String]("prgseed").optional().valueName("long").
       action((x, c) => c.copy(ldaPRGSeed = Some(x.toLong))).
       text("seed for the pseudorandom generator")
-*/
+
     opt[Int]("ldaMaxIterations").optional().valueName("int").
       action((x, c) => c.copy(ldaMaxIterations = x)).
       text("maximum number of iterations used in LDA")
+
+
+    opt[Double]("ldaAlpha").optional().valueName("float64").
+      action((x, c) => c.copy(ldaAlpha = x)).
+      text("document concentration for lda, default 1.02")
+
+    opt[Double]("ldaBeta").optional().valueName("float64").
+      action((x, c) => c.copy(ldaBeta = x)).
+      text("topic concentration for lda, default 1.001")
   }
 }
